Still trying to fix the client

+++ src/main/java/pl/pjtom/Client.java
@@ -29,9 +29,10 @@ public class Client implements Runnable {
             ArrayList<PackageModel> packagesForMe = cassClient.getPackagesInPostBoxByClientID(postBox.getPostBoxID(), clientModel.getClientID());
 
             for (PackageModel p: packagesForMe) {
-                if (p.getIsReadyToPickup()) {
+                if (!p.getWasPickedUp() && p.getIsReadyToPickup()) {
                     Date timestamp = new Date(System.currentTimeMillis());
                     p.setIsReadyToPickup(false);
+                    cassClient.updatePackageWasPickedUpFromPostBox(postBox.getPostBoxID(), p.getPackageID());
                     cassClient.deletePackageFromPostBox(postBox.getPostBoxID(), p.getPackageID());
                     cassClient.upsertPackageLog(new PackageLogEntryModel(p.getPackageID(), PackageLogEvent.PICKUP_PACKAGE_FROM_POSTBOX, timestamp, clientModel.getClientID(), p.getPostBoxID()));
                     System.out.println(clientModel.getClientID() + ": Taking package " + p.getPackageID() + " from postbox " + postBox.getPostBoxID());
